Make high rate tests sleep
Because the loop did not have any sleep value in it, the token bucket was receiving requests on the order of microseconds, causing it to treat the incoming requests as a "burst".

We use a default (unchangeable) burst value of 1, as we are only using the `Allow()` method to allocate one token per request, rather than requesting multiple tokens at once.