We received a vulnerability disclosure due to how we pick a remote IP address.
Disclosure URL: https://gist.github.com/adam-p/4b777de4bda0027f4c3daa45618adcdc

This is an attempt to address the situation.

1. We no longer configure SetIPLookups on default.

2. We address the two different SetIPLookups confusion in two different place by removing both of them.

3. We add a new, explicit way, for user to define how IP address should be picked up.

Tests are all updated to use the new method of picking IP address.

This will be a backward incompatible change so version number has to be bumped to 7.